% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlim.R
\docType{class}
\name{jlim-class}
\alias{jlim-class}
\title{jlim class to contain the processed input data and results of JLIM test}
\description{
jlim class to contain the processed input data and results of JLIM test
}
\section{Slots}{

\describe{
\item{\code{assoc1}}{data.frame of trait 1 with markers acutally used}

\item{\code{assoc2}}{data.frame of trait 2 with markers acutally used}

\item{\code{ld0}}{reference LD of markers actually used}

\item{\code{ld2}}{in-sample LD of trait 2 cohort only with markers actually used}

\item{\code{idx2BP}}{BP position of most associated SNP in assoc1}

\item{\code{STAT}}{jlim statistic lambda}

\item{\code{p.value}}{permutation p-value}
}}

