% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlim_gencfg.R
\name{jlim.gencfg}
\alias{jlim.gencfg}
\title{command-line interface to generate config file.}
\usage{
jlim.gencfg()
}
\description{
This function scans through index SNPs listed in provided indexSNP file, 
and lists up all trait1-trait2 combinations to examine later by
jlim.test(). Typically, trait 1 is GWAS association and trait 2 is eQTLs.
Regarding the file formats, please see accompanied example test cases 
in the distribution.
}
\details{
Usage:
\code{jlim_gencfg.sh}

Command line arguments:
\itemize{
\item \code{--tr1-name}: trait 1 name (no white space allowed)
\item \code{--tr1-dir}: directory containing trait 1 association summary statistics
\item \code{--idxSNP-file}: indexSNP file
\item \code{--tr2-dir}: directory containing trait 2 association data
\item \code{--refld-dir}: reference LD panel (extracted by fecth.refld0.EUR.sh)
\item \code{--out}: output config file name
\item \code{--p-tr2-cutoff}: trait 2 association p-value cut-off to run the
analysis (default 0.01)
\item \code{--reassign-window}: add this option to readjust analysis window when provided index SNPs are not the most associated SNPs in trait 1 association summary data.
}

\code{--tr1-name --tr1-dir --tr2-dir --idxSNP-file --refld-dir} are required
arguments, and the rest are optional.
}

