% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlimCmd.R
\name{run.jlim}
\alias{run.jlim}
\title{Joint Likelihood Mapping (JLIM) test on a single locus}
\usage{
run.jlim()
}
\description{
command-line interface to to run JLIM test on a single locus.
JLIM tests whether two traits - main and secondary - are driven by shared
causal effect or not. The method is described in Chun et al. (
http://biorxiv.org/content/early/2016/05/12/053165)
}
\details{
Command line arguments:
\itemize{
\item \code{--maintr-file}: main trait association summary statistics file names including path.
\item \code{--sectr-file}: second trait association summary statistics file.
\item \code{--maintr-ld-file}: main trait ld file contains genotype or pre-calculated ld-matrix of the first cohort.
 It has to contain identical variants as the main trait association file. currently genotype file could be of .ped or .hap file.
\item \code{--maintr-ld-format}: in case maintr-ld-file  is filled with a ldmatrix file name, this argument should be assigned with ldmatrix.
\item \code{--sectr-ld-file}: second trait ld file contains genotype  of the second cohort.
 It has to contain identical variants as the second trait association file.
directory containing trait 2 association data
\item \code{--ref-ld}: a directory address that contains the complete reference panels (already constructed for the non-Finish European from  1000 genome). It could also be an already chosen reference panel that covers the locus.
\item \code{--index-snp}: this option accept the value in chr:BP format for index SNP.
Based on the specified BP, JLIM calculate an interval for its test.
\item \code{--manual-window-boundary}: if provided JLIM will override and readjust it's already calculated interval based on the index SNP.
\item \code{--perm-file}: assigning this parameter JLIM assumed offline option unless explicitly specifying the other format.
\item \code{--max-perm}: in case of on-the-fly permutation, JLIM expect the maximum number of the permutation
\item \code{--sectr-ref-db}: currently JLIM apart from the normal association file specified by the user
it can accept the second trait association file from two different file source (eQTL catalog or GTEx data).
in case user likes to test his primary trait against eQTL data catalog, he has to assign eQTL to this parameter.
for GTEx data source JLIM can handle it based the file extension.
eQTL has .tsv indexed file format which need .tbi could be read directly read drom it's original file source,
GTEx has .parquat file format which can only be read locally]
JLIM automatically extract and assign the column names of these two formats but to cover different and new data sources, user has to specifies column names in the below arguments.
\item \code{--sectr-colname-variant-id}: in case of reading GTEx .parquat format user has to specify the name of the variant-id column.
\item \code{--sectr-sample-size}: in on-the-fly permutation mode, to increase the accuracy of the results user has to specify the sample size of second cohorts that second association file is produced from.
\item \code{--sectr-gene-filter}: second trait's gene name to filter in case it contains multiple genes.
\item \code{--output-file}: name of the output file
\item \code{--maintr-colname-chr}: main trait's column name which contains chromosome Id
\item \code{--maintr-colname-bp}: main trait's column name which contains base pair position of variant.
\item \code{--maintr-colname-p}: main trait's column name which contains pvalue.
\item \code{--sectr-colname-chr}: second trait's column name which contains chromosome Id
\item \code{--sectr-colname-gene}: second trait's column name which contains gene Id
\item \code{--sectr-colname-bp}: second trait's column name which contains base pair position of variant.
\item \code{--sectr-colname-p}: second trait's column name which contains pvalue.
\item \code{--r2-resolution}: r2 resolution limit of JLIM analysis (default r2=0.8)
\item \code{--window-size}: the locus size to run JLIM test (default 200k)
\item \code{--min-MAF}: minimum value of minor allele frequency for the second traits and reference panel(default  0.05)
\item \code{--min-SNPs-count}: minimum number of the common SNPs in the main and second traits to run JLIM test (default 50).
\item \code{--min-pvalue}: pvalue minimum threshold for second trait association pvalue to run the
analysis (default 0.05)
\item \code{--save-to-rda-file}: path and prefix of the file name to save JLIM objects in the rda files
\item \code{--save-to-rda-p}: minimum required pvalue of the JLIM result in order to save the data objects into the output files
\item \code{--recessive-model}: if it is not null JLIM will consider the recessive SNPs inside the main trait and run accordingly
}

\code{--maintr-file --sectr-file  --index-snp} are required
arguments, and the rest of the arguments could be optional or mandatory based on the requstedpermutation mode.
}
