% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlim_cmdutil.R
\name{run.jlim}
\alias{run.jlim}
\title{command-line interface to \code{\link{jlim.test}}}
\usage{
run.jlim()
}
\description{
See accompanied run_jlim.sh to execute this function.
}
\details{
Usage:
\code{run_jlim.sh input.cfg.tsv r2_resolution outfile}

Command line arguments:
\itemize{
\item \code{input.cfg.tsv}: config file generated by \code{\link{jlim.gencfg}}
\item \code{r2_resolution}: r2 resolution between 0.0 and 1.0 (default 0.8)
\item \code{outfile}: output file. Table with two column STAT and p. STAT is JLIM statistic lambda. p is the permutation p-value.
}
}

